+++ src/lib.rs
@@ -416,7 +416,6 @@ pub mod opaque;
 pub mod ciphersuite;
 mod envelope;
 pub mod hash;
-mod serialization;
 
 mod elligator;
 pub mod group;
@@ -432,6 +431,8 @@ pub mod oprf;
 mod oprf;
 
 pub mod slow_hash;
+
+mod serialization;
 
 #[cfg(test)]
 mod tests;
